requirements.txt
absl-py==1.1.0
cachetools==5.2.0
certifi==2021.5.30
charset-normalizer==2.1.0
cycler==0.11.0
filelock==3.7.1
fonttools==4.33.3
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.19.3
joblib==1.2.0
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.5.2
mkl-service==2.4.0
networkx==2.6.3
numpy
oauthlib==3.2.1
olefile==0.46
packaging==21.3
Pillow==9.0.0
progress==1.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.5.4
tokenizers==0.12.1
torchvision>=0.13.0
tqdm==4.64.0
transformers==4.20.1
typing-extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
pynacl
wandb
aim
